+++ src/protolink/__init__.py
@@ -1,4 +1,4 @@
 """ProtoLink package."""
 
-__version__ = "0.2.4"
+__version__ = "0.2.5"